refactor(mobile): Move the logic of the integration getter in the mediator
The logic was in the repository and was returning a Future<List<String>> instead of a Future<Response<?>>  like every repository
